Fix covariance overparameterisation (#150)
GPflow often optimizes positive-definite matrices. To maintain positive-definiteness without constrained optimization, a lower-triangular matrix is optimized. 

  Sigma + L L ^T

The previous approach to optimizing L was to ignore the upper half. The mean  that there were some extra variables in the optimization vector, which did nothing. This PR implements a tensorflow op which transforms back-and-forth between triangular matrix L and a 'packed' vector representation. The result is that there are no redundant parameters in the optimization vector.