Keep AutoFlow computations off the default graph
The Autoflow function now makes its own graph for each autoflow in the
model. Some extra tests in test_AutoFlow are added to make sure that it
works with dataholders.

The dataHolder class has changed a little: it's included in
make_tf_array now, so that the placeholders can be made inside the
Autoflow graph: before, all placeholders were on the default graph.