Refactor of EP implementation to modularise and make more extensible in the future.
* EP while loop broken out and made into seperate class to allow it to
be easy to reuse elsewhere. Tidy up of the variable names and made more
general. Tests written to check that it can recover the true solution
when the approximating factors have the same solution as the true
solution.
* Central moment functions moved out of EP class and live elsewhere.
Step function, Gaussian and Gaussian CDF moments implemented all with
tests.
* Likelihoods now are associated with a method that returns their central
moments generator function. Only implemented for Bernoulli but this
allows one to store moment logic with the likelihood function which is
more where it belongs.
* Updated the VGP versus EP binary classification example script to use
this new source structure. This has remained in the docs so have not
written tests for this part.