Likelihood/students t variance scaling (#777)
* Add scaling to studentT conditional variance

The conditional variance of the Student’s T distributions is
proportional to the square of the scale of the distribution.  See
https://en.wikipedia.org/wiki/Student%27s_t-distribution#In_terms_of_sca
ling_parameter_σ,_or_σ2.
I’ve incorporated the correct scaling factor.

* explicit scale dtype and tensor broadcasting

Added an explicit data type for the Student’s T scale parameter, and
made the broadcasting in the conditional_variance method explicit.