Cleaning up stationary kernel implementations (#827)
Stationary kernels are now implemented by K_r2 as a function of the scaled square distance or K_r as a function of the scaled Euclidean distance, see the docstrings for details. scaled_square_dist and scaled_euclid_dist are deprecated, and square_dist and euclid_dist are removed. RBF is now an alias for SquaredExponential.