Fix AutoBuild metaclass to not screw up method signatures (#907)
The AutoBuild metaclass that is present in all Parameterized subclasses (e.g. models, kernels, likelihoods etc.) overwrote the `__init__` function and thereby replaced the signature with `*args, **kwargs`, which meant that the API documentation and command-line/IDE auto-completion was broken. This PR makes use of the "decorator" package (slightly hacked) to restore the original signatures (while allowing autobuild to still be passed as an additional keyword argument).