fix bug in log_prior when a Parameter has a transform but no prior (#1099)
Fixes a bug in the optimization objective for all `BayesianModel` subclasses when the model has parameters that have a `transform` (e.g. `positive()`, for kernel lengthscales and variances) but no `prior` (i.e. when doing Maximum Likelihood Estimation (MLE) inference). The objective should be independent of the parametrisation of the parameters, i.e. `log_prior()` of the parameters should be zero.