Fix hidden bug in SGPR
SGPR likelihoods were previously using full rank matrices instead of
diagonal ones in both upper bound and likelihood calculation. Ie `Kdiag`
was not "diag".

This error was being masked by the intentional deactivation of tests
comparing to the SGPR to the GPR, and what appears to be a hack to make
tests working on the upper bound case.

Both are fixed here, and tests reactivated.