Added the sgpr method sgpr.compute_qu, from gpflow1, and fixed test
The method sgpr.compute_qu had not been migrated from gpflow1, and a
test that was meant to check it had been patched up to pass,
erroneously.

After speaking to @markvdw, concluded this method is useful, in
particular to compare to SVGP model. The test has been patched up and
the method ported to gpflow2.