Fix hidden bug in SGPR (#1106)
1. Fix hidden bug in SGPR
2. Add the  sgpr.compute_qu method from gpflow1

1. [Bug]. SGPR likelihoods were previously using full rank matrices instead of
diagonal ones in both upper bound and likelihood calculation. Ie `Kdiag`
was not "diag". 

This error was being masked by the intentional deactivation of tests
comparing to the SGPR to the GPR, and what appears to be a hack to make
tests working on the upper bound case.

2. [Migration]. Fixing the above broke another test, originally used for
 sgpr.compute_qu.  The method sgpr.compute_qu had not been migrated 
from gpflow1, and a test that was meant to check it had been patched up to pass,
erroneously.

After speaking to @markvdw, concluded this method is useful, in
particular to compare to SVGP model. The test has been patched up and
the method ported to gpflow2.