Clean up `presliced` argument to kernel calls (#1235)
In GPflow 1, each kernel's K and K_diag method was supposed to take a presliced argument. Not all kernels actually had this signature, and not all kernels who had this signature actually did anything with the argument. This PR makes presliced an argument solely to the kernel's __call__, with the intent of making the entire set-up a lot cleaner.