remove trainable setter on Parameter (#1323)
TensorFlow does not allow setting a tf.Variable's `trainable` property; this brings GPflow Parameter objects in line with that behaviour. gpflow.set_utilities() can be used to change the _trainable state for gpflow.Parameter and tf.Variable objects as well as recursively for tf/gpflow Modules.

Note that set_trainable() should not be used inside functions that will be wrapped in tf.function() (see #1335).