Posterior object for fast predictions (#1636)
GPflow had already been optimised for speed in the training loop. This PR addresses the speed of predictions and resolves (partially) #1599. We achieve this by introducing new Posterior objects that can cache the results of a "precompute" stage (everything that's independent of test points) and thereby significantly speeds up repeated predictions.

Co-authored-by: ST John <st--@users.noreply.github.com>
Co-authored-by: stefanos <stefanosele@users.no-reply.github.com>
Co-authored-by: John Mcleod <43960404+johnamcleod@users.noreply.github.com>